+++ pkg/acceptance/bettertestspoc/README.md
@@ -351,3 +351,4 @@ func (w *WarehouseDatasourceShowOutputAssert) IsEmpty() {
     1. Lists of objects are partially generated, and only parameter name is generated in some functions (the type has to be added manually).
     2. `testing` is a package name that makes Go think that we want to have unnamed parameter there, but we just didn't generate the type for that field in the function argument.
 - generate assertions checking that time is not empty - we often do not compare time fields by value, but check if they are set
+- support generating provider config and use generated configs in `pkg/provider/provider_acceptance_test.go`
